Move livestroke into separate file
