+++ src/board/stroke/shape.tsx
@@ -11,28 +11,8 @@ import {
     SEL_STROKE,
     SEL_STROKE_ENABLED,
 } from "consts"
-import store from "redux/store"
 import { useCustomSelector } from "redux/hooks"
 import { LiveStroke, Point, Stroke, ToolType } from "drawing/stroke/types"
-
-export const LiveStrokeShape = memo(() => {
-    useCustomSelector((state) => state.drawing.liveStrokeUpdate) // trigger update
-    const { liveStroke } = store.getState().drawing
-    return (
-        <>
-            {liveStroke.pointsSegments.map((subPts: number[], i: number) => {
-                const strokeSegment = {
-                    ...liveStroke,
-                    points: subPts.slice(),
-                } as LiveStroke
-                // we can use the array index as key here
-                // since the array's order is not changed
-                // eslint-disable-next-line react/no-array-index-key
-                return <StrokeShape key={i} stroke={strokeSegment} />
-            })}
-        </>
-    )
-})
 
 interface StrokeShapeProps {
     stroke: Stroke | LiveStroke
